Add Edna.Dashboard.DB.Schema
Problem: we tend to split DB schema into multiple modules, but some
tables are still in the main Edna.DB.Schema which is supposed to only
group existing tables from other components.

Solution: create Edna.Dashboard.DB.Schema module and move tables there.
