+++ backend/edna.cabal
@@ -33,6 +33,7 @@ library
   exposed-modules:
       Edna.Config.Definition
       Edna.Config.Utils
+      Edna.Dashboard.DB.Schema
       Edna.Dashboard.Web.API
       Edna.Dashboard.Web.Types
       Edna.DB.Connection
